[PRMDR-471]
Most of the way refactored, but my laptop is about to shut down for an update
