+++ lambdas/utils/lloyd_george_validator.py
@@ -7,8 +7,9 @@ from enums.pds_ssm_parameters import SSMParameter
 from enums.supported_document_types import SupportedDocumentTypes
 from models.nhs_document_reference import NHSDocumentReference
 from models.pds_models import Patient
-from pydantic import ValidationError
 from requests import HTTPError
+
+from models.staging_metadata import StagingMetadata
 from services.document_service import DocumentService
 from services.ssm_service import SSMService
 from utils.audit_logging_setup import LoggingService
@@ -24,10 +25,13 @@ class LGInvalidFilesException(Exception):
     pass
 
 
+file_name_invalid = "One or more of the files do not match the required file type"
+
+
 def validate_lg_file_type(file_type: str):
     if file_type != "application/pdf":
         raise LGInvalidFilesException(
-            "One or more of the files do not match the required file type"
+            file_name_invalid
         )
 
 
@@ -36,7 +40,7 @@ def validate_file_name(name: str):
     lg_regex = rf"[0-9]+of[0-9]+_Lloyd_George_Record_\[{REGEX_PATIENT_NAME_PATTERN}\]_\[{nhs_number_pattern}\]_\[\d\d-\d\d-\d\d\d\d].pdf"
     if not re.fullmatch(lg_regex, name):
         raise LGInvalidFilesException(
-            "One or more of the files do not match naming convention"
+            file_name_invalid
         )
 
 
@@ -58,7 +62,7 @@ def check_for_number_of_files_match_expected(file_name: str, total_files_number:
             )
     except (AttributeError, IndexError, ValueError):
         raise LGInvalidFilesException(
-            "One or more of the files do not match naming convention"
+            file_name_invalid
         )
 
 
@@ -96,6 +100,16 @@ def validate_lg_file_names(file_name_list: list[str], nhs_number: str):
     validate_with_pds_service(file_name_list, nhs_number)
 
 
+def validate_bulk_uploaded_files(staging_metadata: StagingMetadata):
+    # Delegate to lloyd_george_validator service
+    # Expect LGInvalidFilesException to be raised when validation fails
+    file_names = [
+        os.path.basename(metadata.file_path) for metadata in staging_metadata.files
+    ]
+
+    validate_lg_file_names(file_names, staging_metadata.nhs_number)
+
+
 def extract_info_from_filename(filename: str) -> dict:
     page_number = r"(?P<page_no>[1-9][0-9]*)"
     total_page_number = r"(?P<total_page_no>[1-9][0-9]*)"
@@ -110,13 +124,13 @@ def extract_info_from_filename(filename: str) -> dict:
         return match.groupdict()
     else:
         raise LGInvalidFilesException(
-            "One or more of the files do not match naming convention"
+            file_name_invalid
         )
 
 
 def check_for_file_names_agrees_with_each_other(file_name_list: list[str]):
     expected_common_part = [
-        file_name[file_name.index("of") :] for file_name in file_name_list
+        file_name[file_name.index("of"):] for file_name in file_name_list
     ]
     if len(set(expected_common_part)) != 1:
         raise LGInvalidFilesException("File names does not match with each other")
@@ -152,9 +166,9 @@ def validate_with_pds_service(file_name_list: list[str], nhs_number: str):
         if patient_details.birth_date != date_of_birth:
             raise LGInvalidFilesException("Patient DoB does not match our records")
         patient_full_name = (
-            " ".join([name for name in patient_details.given_Name])
-            + " "
-            + patient_details.family_name
+                " ".join([name for name in patient_details.given_Name])
+                + " "
+                + patient_details.family_name
         )
         logger.info("Verifying patient name against the record in PDS...")
 
@@ -165,7 +179,7 @@ def validate_with_pds_service(file_name_list: list[str], nhs_number: str):
         if patient_details.general_practice_ods != current_user_ods:
             raise LGInvalidFilesException("User is not allowed to access patient")
 
-    except (ValidationError, ClientError, ValueError) as e:
+    except (ClientError, ValueError) as e:
         logger.error(e)
         raise LGInvalidFilesException(e)
     except HTTPError as e:
